+++ examples/mediation/docker-compose.yml
@@ -95,7 +95,7 @@ services:
       - BOB=http://bob:3001
       - MEDIATOR=http://mediator:3001
     volumes:
-      - ../../controller:/usr/src/app/controller:z
+      - ../../acapy_controller:/usr/src/app/acapy_controller:z
       - ./example.py:/usr/src/app/example.py:ro,z
     command: python -m example
     depends_on:
